+++ client/src/config/define.js
@@ -1,0 +1,28 @@
+// ドメイン
+const APP_DOMAIN = 'coodig.com'
+// ローカルIPアドレス
+const LOCAL_IP = '192.168.33.11';
+// URL(IP)
+const FULL_LOCAL_URL = 'http://192.168.33.11';
+// URL(ドメイン)
+const FULL_URL = 'http://coodig.com';
+// URL(ドメインwww有り)
+const FULL_URL_WWW = 'http://www.coodig.com';
+// サーバーサイドポート番号
+const SERVER_PORT = 5000;
+// フロントサイドポート番号
+const FRONT_PORT = 3000;
+// トップに出すPGタグの数
+const TOP_VIEW_TOTAL_NUMBER = 10;
+
+export default{
+    APP_DOMAIN:APP_DOMAIN,
+    LOCAL_IP:LOCAL_IP,
+    FULL_LOCAL_URL:FULL_LOCAL_URL,
+    FULL_URL:FULL_URL,
+    FULL_URL_WWW:FULL_URL_WWW,
+    SERVER_PORT:SERVER_PORT,
+    FRONT_PORT,FRONT_PORT,
+    TOP_VIEW_TOTAL_NUMBER: TOP_VIEW_TOTAL_NUMBER,
+};
+
