+++ server/API/api_getPlangColor.js
@@ -1,9 +1,9 @@
 //======================================================================
 // set up
 //======================================================================
-const dbConf = require('../dbManager');
-const dbQuery = require('../../lib/db_query');
-const define = require('../../config/define');
+const dbConf = require('../DbManager');
+const dbQuery = require('../lib/db_query');
+const df = require('../config/define');
 //======================================================================
 exports.getAPI = function (req,res) {
     const SELECT_ALL_REACT_TEST_Q = 'SELECT * FROM p_lang_mst';
