+++ server/DbManager.js
@@ -1,5 +1,5 @@
 const mysql = require('mysql');
-const dbconfig = require('../config/db_conf');
+const dbconfig = require('./config/db_conf');
 // DB接続
 const connection = mysql.createConnection(dbconfig.connection);
 // DB設定情報出力
