+++ server/lib/db_query.js
